+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 BREAKING CHANGES:
 
 - [DeliverTx] Check node ID is valid and role is `AS` (`RegisterServiceDestinationByNDID`, `EnableServiceDestinationByNDID` and `DisableServiceDestinationByNDID`).
+- [DeliverTx] Add new functions (`RevokeAccessorMethod`).
+- [DeliverTx] Check request is closed with valid ial, proof and signature (`AddAccessorMethod`).
 
 ## 0.10.1 (October 9, 2018)